// src/app/model/form-values.model.ts
export interface FormValuesModel {
  name: string;
  isOld: string;
  isYoung: string;
  phoneNumber: string;
  country: string;
  province: string;
  trump: string;
}

export interface UserInfoModel {
  id: number;
  FirstName: string;
  LoserName: string;
  Age: string;
  Hotline: string;
}

export interface UserLocationModel {
  Country: string;
  ProVince: string;
}